Migrate declaration of `Message` data class
